task3

+++ controller/post.js
@@ -1,9 +1,10 @@
 const Post = require("../models/post.js");
-const createError = require("../middleware/error.js");
+const { createError } = require("../middleware/error.js");
 
 // create a new post
 const addPost = async (req, res, next) => {
-    const newPost = new Post({ userId: req.user_id, ...req.body });
+    console.log("user id of the request", req.user.id)
+    const newPost = new Post({ userId: req.user.id, ...req.body });
     try {
         const savePost = await newPost.save();
         res.status(200).json(savePost);
@@ -12,21 +13,26 @@ const addPost = async (req, res, next) => {
     }
 };
 
+
 // update a post
 const updatePost = async (req, res, next) => {
     try {
         const post = await Post.findById(req.params.id);
         if (!post) return next(createError(404, "Post not found"))
-        const updatedpost = await Post.findByIdAndUpdate(req.params.id,
-            {
-                $set: req.body,
-            },
-            { new: true }
-        );
+        if (req.user.id === post.userId) {
+            const updatedpost = await Post.findByIdAndUpdate(req.params.id,
+                {
+                    $set: req.body,
+                },
+                { new: true }
+            );
+            res.status(200).json(updatedpost)
 
-        res.status(200).json(updatedpost);
+        } else {
+            res.status(403).json("You can only update your post")
+        }
     } catch (err) {
-        next(err);
+        res.status(400).json({error:err.message})
     }
 };
 
@@ -35,31 +41,34 @@ const deletePost = async (req, res, next) => {
     try {
         const post = await Post.findById(req.params.id);
         if (!post) return next(createError(404, "post not found"))
+        if (req.user.id === post.userId) {
 
-        await Post.findByIdAndDelete(req.params.id);
-
-        res.status(200).json("post is deleted");
+            await Post.findByIdAndDelete(req.params.id);
+            res.status(200).json("post is deleted");
+        } else {
+            res.status(403).json({message:"You can only delete your post..sign in to diff acct "})
+        }
     } catch (err) {
-        next(err);
+        res.status(500).json({error:err.message})
     }
 };
 
 // get a post by id
-const getPost = async (req, res, next) => {
+const getPost = async (req, res) => {
     try {
         const post = await Post.findById(req.params.id);
         res.status(200).json(post)
     } catch (err) {
-        next(err);
+        res.status(500).json({error:err.message})
     }
 };
 //get all posts
-const getAllPosts = async (req, res, next) => {
+const getAllPosts = async (req, res) => {
     try {
         const posts = await Post.aggregate([{ $sample: { size: 20 } }]);
         res.status(200).json(posts)
     } catch (err) {
-        next(err);
+        res.status(500).json({error:err.message})
     }
 };
 
@@ -71,7 +80,7 @@ const likePost = async (req, res, next) => {
         })
         res.status(200).json("The post is liked")
     } catch (err) {
-        next(err)
+        res.status(500).json({error:err.message})
     }
 }
 
@@ -84,10 +93,9 @@ const commentPost = async (req, res, next) => {
         await Post.findByIdAndUpdate(req.params.id, {
             $push: { comments: req.body.comment }
         });
-
         res.status(200).json({ message: "Comment added successfully" });
     } catch (err) {
-        next(err);
+        res.status(500).json({error:err.message})
     }
 };
